Summary: 	Added DebugLib support. 	Fixed memory handling.
Detail:
	Altered memory handling so that the DA size is not increased *every*
time some more memory is required, but only if the DA has run out of space.
This saves memory and means that we can download a Funai 8 image. (Note this
only failed if the server did not return the size of the file, and hence the
memory to hold the file had to be dynamically allocated as the image was
being downloaded).
	Added DebugLib support to work out why the download was not working.
	Tidied up the code so that memory that was being malloc()'ed is now
free()'d when it is no longer required.

Admin:
	Tested. Which was more than the original version was ;-) The code
could really do with being tidies, however.

Version 0.03. Tagged as 'UpgFetch-0_03'