Summary: 	This program now parses an upgrade file in the form specified in the specification (D.O. number currently unallocated) and flashes the image held within.
Detail:
	The program reads the upgrade file into a dynamic area. As it streams
the file into the DA, it splits the data into Tag-Length-Value triplets and
stores these triplets in a linked list, with the value pointing to the
location of the data within the Dynamic Area.
	Some of these TLV tags are acted upon during the file download (e.g.
the file length TLV tag sets the size of the DA to the size passed in),
whilst others are acted upon once the download has been completed (e.g. the
CRC check TLV tag).
	In this version of the program the streaming of the file works, and
the TLV decode is supported. Currently the only two TLV's acted upon are the
file size tag (used to set the size of the Dynamic Area in which to store the
file) and the display string tag (used to display a string during download).
	A download of an image file from the intranet took about 10 seconds
on an Ethernet Funai 8 NC, and about 10 minutes on a Modem Funai 8 NC.
Decompressing and flashing of the image takes under a minute.

Admin:
	Work in progress. Support for more TLV's still need to be added.
	Tested fairly extensively on both Ethernet and Modem NC's.

Version 0.05. Tagged as 'UpgFetch-0_05'