Summary: 	Altered the MakeImage utility.
Detail:
	Now supports CSFS tags being available within the configuration file.
To compile this functionality in, then compile the code with the SUPPORT_CSFS
preprocessor flag set.
	Altered the DefaultCountry and ModemCountry tags so that their types
are now correct. This will stop a potential bug from occuring.
	Tags are now written into the BootBlock area if they have a TagType
greater than zero. This allows us to disperse the tags in the internal
big_list in a more sane manner.
 	Also added an example config file for CSFS support.

Admin:
	Tested.
	Please note that if giving this code to outside customers for the
Alba-based projects they must not get any of the SUPPORT_CSFS-enabled code.
Plase strip this out if sending the code outside the company.

Version 0.50. Tagged as 'NVRAM-0_50'