Summary: 	Added a Library directory into !Boot
Detail:
	Fumai want ping and traceroute to be available within the ROM.
Unfortunately doing this the traditional way (by including
SystemRes/Internet) in te build leads to the ROM size being to big as it
builds *all* the utilities. Therefore have done a binary import of ping and
traceroute into <boot$dir>.Library, and set the run$path to point to it.

Admin:
	Untested. Should work, hopefully, with a following wind and some
luck.

Version 1.04. Tagged as 'Funai3UI-1_04'