Summary: 	Initial work to get CSFS (Smurfs) working in NCMALite.
Detail:
	CSFS allows FTP and HTTP mounts to be treated as a local filing
system, allowing programs to be executed off servers.
	It is a requirement for the Bethany project for testers to be able to
load and execute test scripts off a server, logging the results on the
server. This work enables this to occur.
	For security and change control reasons, we do not wish to have the
CSFS module in the ROM chain. For this reason on network initialisation NCMA
checks for CSFS's presecnce and if a certain bootblock tag (CSFS_TYPERO) is
set then it attempts to softload CSFS out of a location in ResourceFS.
	If all the CSFS bootblock tags have been set correctly, then NCMA
attempts to create the CSFS mounts once a dial has succesfully completed. If
the mount(s) work, then it attempts to run a boot file off the server (once
each for read-only and read/write mounts).
	There are two types of CSFS mounts available within NCMA - read only
and read/write. Both of these are supported (and have been tested) in this
version.
	This new functionality is currently disabled in this version. To
enable it, then set the SOFTLOAD_CSFS, ALLOW_CSFS_BOOTING and
WRITE_THRU_SMURFS preprocessor flags in options.h. These flags are currently
commented out.
	Also changed references to BOOL to bool (as defined in <stdbool.h>),
so that DebugLib now works.

Admin:
	Work in progress. This version connects up to my local user area with
bith read-only and read-write mounts. It can execute a program and save data
to the user area using WebFTP as the fetching service.
	No dismounting is currently done; when the line drops CSFS should be
told that the mouts are no longer available. It also needs more testing with
(for example) HTTP fetchers.
	This work also needs merging onto the trunk when it is completed.

Version 0.72, 1.75.2.1. Tagged as 'NCMAlite-0_72-1_75_2_1'