Summary:    Initial support for fetching via proxies added.
Detail:
   Funai have requested that the upgrade system will be able to fecth the
upgrade image via a proxy if one is configured. This image contains the
initial work to enable this to happen. The proxy server details are retrieved
from the bootblock.
   The setup_proxy() routine may need a call to URL_Parse to be made to
ensure that the proxy server retrieved from the bootblock is in the correct
format for the URL module.

Admin:
   Work in progress. This code has had minimal testing.

Version 0.14. Tagged as 'UpgFetch-0_14'