Summary:     Now allows binary to be built during the resources phase.
Detail:
    For years IZipFS and PPADriverJ have been binary exports in the Boot
component. Now we are moving to a 32-bit world, we need to have both 26 and
32 bit compatible versions. Rather than have two binaries in CVS (and have
the normal hassles of keeping them up to date), PPADriverJ has been altered so that if you pass in the flags
      -options TYPE=install RES=!Boot.Modules it will install the built
module to the given directory during the resources phase. If there is no TYPE
flag then the component will build as normal.

Admin:
    Tested in a Lazarus 32-bit build.

Version 1.00, 1.2.2.3. Tagged as 'PPADriverJ-1_00-1_2_2_3'