Summary:     Improved memory handling.
Detail:
    The previous versions of ISPCapture used an ancient version of the
dynalloc.c routines (which are also used in the NCMA* stacks). That version
always uses dynamic areas to store data on modern versions of the OS - so if
you wanted to store 20 bytes it would use an entire page of memory.
    I have replaced the ancient version with an up-to-date version out of
NCMAlite. This version only uses a dynamic area to store data if the required
size is above a certain threshold and otherwise uses malloc().

Admin:
    Built and appears to work correctly. Since this is a drop-in replacement
from NCMAlite it should be fairly stable. Famous last words ;-)

Version 0.28. Tagged as 'ISPCapture-0_28'