Summary:   Fixed to work at startup and when interrupts are off for a while.
Detail:
  The module now resets the USB port during initialisation so that the keyboard
gets detected.

  The notification bit for other interrupts is now cleared from the write back
done head pointer so that it correctly matches the physical address of the trans
fer descriptor.  This was preventing operation after interrupts were off for a w
hile as other interrupts would then be set.

  Keys are now explicitly unpressed when the module is killed so as not to
leave the machine repeating keypresses.  The mouse pointer type is also
reverted to the previous type.

Admin:
  Tested on the LinkUp SDB