Summary:   Fixed bug in Exit handler
Detail:
  The new addition which allows an exit code to be passed using the syntax
"QUIT <expression>" was incorrectly assuming it had access to the workspace
via ARGP when it didn't, this caused a data abort when issuing *quit from
BASIC, or in fact when any program called from BASIC called it's exit
handler.

  It turned out that the required parameters for OS_Exit were already being
pointed to by R12, so a simple LDMIA was sufficient.

Admin:
  Tested on a RiscPC



Version 1.25. Tagged as 'BASIC-1_25'