Summary:     Now works.
Detail:
    The version of Stewart's notes that I was using to convert the component
to support ObjAsm was out of date, and hence missed out a couple of sections
on converting Asm headers to C headers and tokenisation. This would only have
been picked up when building straight out of CVS. A classic case of RTLM
(read the latest manual). However, doing this work showed up a bit missing
from Stewart's notes, so it is really a case of RTLMWSHWI, or read the latest
manual when Stewart has written it). ;-)
    The makefile has been altered to ensure that the correct operations occur.

Admin:
    Not tested in a build. Checked the logs from running export headers,
libs, resources and rom phases against a pre-objasm build and verified that
all the same steps are occuring.

Version 2.58. Tagged as 'FileSwitch-2_58'