Summary:     New ncma_urc added.
Detail:
    The ncma_urc RegistrationRequest has been added. Calling this causes
NCMAlite to attempt a dial using the telephone number stored in the bootblock
tag Registration_PSTN_Num. No validation of the phone number is performed.
    Also removed an errant throwback line from the makefile which was causing
throback to occur in ROM builds (DOH!), and a couple of warnings from the
build caused by routines not being declared in header files.

Admin:
    Tested in a Lazarus build.

Version 0.85. Tagged as 'NCMAlite-0_85'