Summary:     Initial import of the Configuration Upgrade Module.
Detail:
    Import of the module with basic functionality present. Things remaining
to be done are outlined in the Docs.ToDo file. This version can accept a
correctly-formatted update file and reset the unit's configuration to the
values held within the file. It can also write the unit's current
configuration to a file.
    The Resources directory is currently filled with some test programs that
get placed into Resources.ConfigUpd during the resources phase of the build.
These will eventually be moved into the Test directory.
    Full DebugLib support present.

Admin:
    This component relies heavilly on NCMALib.
    Tested in a Lazrus build.
    Work in progress.

Version 0.01. Tagged as 'ConfigUpd-0_01'