Summary:     Slight code rationalisation.     Incremental updates now work.
Detail:
    The routines create_bootblock() display_bootblock() have been moved from
the CU module to the NCMALib library - they are likely to be hevailly used by
other applications.
    Various other support code and preprocessor macros also moved over.
    Incremental update files now work - I knew that I had put most of the
support code in, and it turned out that I only had to alter one line to make
it work.
    Added a file describing the way to use the test programs.
    Now only attempts to write NVRAM tags if the tag has a value. This stops
debug warnings.
    Added a source file c/Validation and a correspondin header file that will
be used to store code  to validate NVRAM and bootblock tags. Thi
functionality may also eventually be moved into NCMALib once it is completed.

Admin:
    Tested in a Lazarus build.

Version 0.02. Tagged as 'ConfigUpd-0_02'