Summary:   Added debugit support.
Detail:
  The debugit module is one of the options for output from Debuglib.  It
buffers output and will set a pollword for an app to display the output at an
opportune time, !ListDebug is the standard handler for this.

  For embedded work it is useful for debug output to come out in the terminal
window on which you are working, since you may not have !Listdebug running on
the client and you probably don't want it to be displayed on the device
anyway.  Issuing the command TServer_DebugItStart will register the
application part of TServer with the DebugIt module, and send debugit output
back across the socket to the telnet terminal being communicated to.

  This provides a capability not available with normal taskwindows - you
can't run a program with debuglib and get output in the taskwindow because
the module producing the output has no idea which taskwindow to output the
debug data to, running from the command prompt you can achieve this as the
machine is then no longer multitasking and you can use printf debuglib
output.  TServer debug output is very similar to this.

  There is a complementary command, TServer_DebugitStop which deregisters
TServer from debugit.

Admin:
  Tested extensively on the LinkUp SDB.  If a lot of debug output is
generated, it can get choked, although it usually seems to come through.
Version 0.03. Tagged as 'TServer-0_03'