Summary:   Fixed memory corruption.
Detail:
  The controller was being put into the operational state before the HCCA had
been set up, so it was writing the frame number (offset &80) to an undefined
location in memory.  The HCCA register was being zeroed upon finalisation, so
the controller was writing to the zero physical page.  This happens to be
screen on the Alba-2 board and was showing up as a white dot then RMReiniting
the module.

  The controller in now not put into operational state until the registers
have been correctly set up, and is put into reset state upon finalisation.

Admin:
  Tested on LinkUp SDB.
Version 0.03. Tagged as 'ProtoUComb-0_03'