Summary:   Fixed Sleep No More, and made 32-bit compatible.
Detail:
  PipeFS uses UpCall 6 (Sleep) to signal that the reading process can sleep,
using the size of the file as a pollword.  Thus when any data is written into
the pipe, the sleeping process will be woken up.  If an attempt is made to
close the file handle, UpCall 7 (Sleep No More) is broadcast to allow anyone
to object to the closure occurring.  Pipe_Close wasn't checking the error
return from making this call, which was an error.  It will now propogate this
error up to fileswitch upon an attempt being made to close the file while a
process is sleeping on it.

  32-bitness involved changing a load of TEQPs for RestPSRs, and fixing some
functions which had forgotten to use [sp, #Proc_RegOffset].

  Also added !Mk and added stripdepend to MkClean.

Admin:
  Tested on LinkUp SDB and RiscPC running a fairly recent Ursula build.
Version 0.21. Tagged as 'PipeFS-0_21'