Summary:   Closing a file now allows the underlying filing system to object, preventing the file from being closed.
Detail:
  The error return from the underlying filing system wasn't being used to
prevent the fileswitch record for the open file being removed.  This resulted
in the underlying filing system not being able to prevent closure of a file
handle.  This behaviour is relied upon by PipeFS when it returns an error
from the Sleep No More UpCall.

Admin:
  Tested on RPC.  Note that the 32 bit build doesn't work on the 26 bit
desktop for no identified reason.
Version 2.60. Tagged as 'FileSwitch-2_60'