Summary:   Improvements for work with debug in interrupt code.
Detail:
  DebugIt can receive data into it's buffer at quite high rates and also
during interrupts when ListDebug has already been paged in by the
Wimp.  The following changes have been made to improve matters:

  1. There is now an Update menu entry for the main window.  If for any
reason, ListDebug has missed some data in the DebugIt data buffer, this will
flush it into the debug window.  It will also act as a 'step' when output is
paused.

  2. Clearing the pollword has now moved to the beginning of the pollword
handler.  Previously it happened after screen update and quite a bit of data
could have come into the DebugIt buffer during that, and would have been
missed if no more came in after the pollword had been cleared (which was the
motivation for change 1).

  3. After reading from the DebugIt buffer in the pollword handler, pollword
nonzero is disabled for one null poll before reenabling it.  This allows the
desktop to continue to function when data is arriving at a high rate.

  4. The background colour of the debug window is now white to avoid the Wimp
tiling it.

Admin:
  Tested on RPC, above features all work and behaviour is enhanced.