NO CODE CHANGES IN THIS COMMIT.
Info only: in yesterday's check-in I mentioned that the fix for runt
pkt error handling appeared to have made little difference.  With a
clear mind this morning I realized that this was because the test
had been invalid.

I'd forced the error handling code to be run by faking the datalength
read from the status block in the controller's buffer.  But since the
packet itself in the controllers' buffer was valid, and the test only
faked the datalength once, it would have recovered and carried on as
normal.  If it really had been a runt packet, the same packet would have
been read again and again, never advancing and therefore never reaching
the exit condition for getting out of the loop.  The fix is valid and
does explain why previously it was possible to get stuck in this loop.

I think there is still the potential for problems in this section of
code, for different reasons (see previous commit).  I'm not convinced
the termination conditions are correct, and indeed looking back to
earlier versions (eg: revision 1.1, there is a patch specific to the
MX (the controller I'm currently using), where the exit test is slightly
different.  I need to investigate some more.


Version 3.18. Tagged as 'EtherI-3_18'