Summary:     Lots of improvements, plus major functionality change.
Detail:
    Added a routine, display_bootblock_tags, that in debug build displays all
the contents of current bootblock (and not NVRAM) tags to the debug stream.
    Many of the 'l' variables now changed to a more meaningful 'loop' to
prevent some of the problems that have been happening with ISPCapture and for
readability of the code.
    The individual bootblock tags now point to a malloc'ed area of memory
containing the value of the tag, and not into the bootblock itself. Whilst
using more memory at runtime this means that individual tags within the
bootblock can be altered and expanded easilly.
    Routines service_configurationchanged(), misc_getenv() and
misc_getenv_malloc() added to ncma_utils as they are used by many things
within the security stack.

Admin:
    Tested with the CU Module on a 32-bit Lazarus build.
    Work in progress.

Version 0.10. Tagged as 'NCMALib-0_10'