Summary:     More work on the exclusion file handling.     Updated test harnesses.     Default timezones are now written back to file correctly.
Detail:
    Initial work on the validation of bootblock data using the rules in the
exclusions file done.
    The NVRAM tag list now also includes an nci_tag-compatible field,
allowing us to have a unique id for all system configuration data.
    There was a bug in which I had totally failed to realise the format in
which the default timezone was written. Hence it was attempting to write it
as a country name rather than a +5:00-style line. This has now been fixed.
    There was another bug inthe handling of the DefaultTerritory NVRAM tag.
This has now been fixed.

Admin:
    Requires NCMALib 0.11
    Run against the latest version of the test harnesses successfully.
    Work in progress, but getting nearer release.

Version 0.05. Tagged as 'ConfigUpd-0_05'