Summary:     A few new routines added.     General tidying up performed.
Detail:
    A new routine, get_line_buffer() added. This allows scanning through an
area of memory to build strings without any length-limited arrays being used.
Memory is dynamically allocated and free'd by the routine free_line_buffer().
This alows no limitation on the length of a string that can be held in a
configuration update file. The amount of memory allocated for the buffer is
increased if the routine detects that more memory is needed for the new
string.
    Another new routine, debug_rmstat() added. This allows current memory
usage to be displayed in debug form with no output appearing via printf().
Consequently the original rmstat() routine has been returned to all printf()
operation.

Admin:
    Tested in a 32-bit Lazarus build.

Version 0.12. Tagged as 'NCMALib-0_12'