Summary:     Exclusions file format changed.     More work on the error handling code.
Detail:
    As work progressed it became obvious that the exclusions file format was
not as flexible as it could have been. for this reason it has been altered so
that it is now <include|exclude> <target> <tag> <rule>. Parsing has also been
improved so that spaces are handled better.
    The error handling code is now present and working. The parsing of the
exclusions and config update files can now generate many errors.

Admin:
    Tested in a 32-bit Lazarus build. Work in progress.
    The FS will be updated to reflect these changes.
    Can be used in conjunction with UpgFetch (versions later than 0.35) to
download and program configuration details into a unit.

Version 0.07. Tagged as 'ConfigUpd-0_07'