Yet another version of SoundDMA. This one's for the LinkUp L7205 or similar, attached to an AC '97 codec that supports Variable Rate PCM Audio.
Tested on the SDB board, with a CS4299 codec. With a bit of luck, should work
on any of the myriad other L7205 systems, as long as the codec reset line is
still on GPIO line PD[0].

With a bit more work, this would form the basis of a generic AC '97 SoundDMA
module. The only nasty wrinkle of the L7205 is that the sound data needs
left-right swapping before playback. Bah.

Version 1.62, 4.8.4.1. Tagged as 'Sound0-1_62-4_8_4_1'