Summary:     Lots of lovely stuff.     Error handling improved.     Test harness functionality expanded and number of cases increased.     Code significatnely tidied up.     New test *command added.
Detail:
    Errors from this module are now returned in the error block. These errors
are in the form specified in the FS, and there can be up to ten errors in the
error string for returning to the server.
    The test harnesses now check the results from the SWI calls and if there
is an error check this error against the expected error. This allows the test
harnesses to test invalid upgrade or exclusion files.
    Some of the code has been shuffled around between various files so that
they are more logically placed - in particular some routines have been moved
into a file with their caller(s) and made static.
    Initial stubs to aid with the validation of NVRAM tags added. Still have
to think up a way of validating some of the NVRAM tags, if it is at all
possible.
    To aid in testing the module a new *command, *cu_change_exc_file, added.
This allows the current exclusions file to be changed, allowing one test
harness to perform many tests. This *command must not be present in any
build that goes into products, and a preprocessor flag will be added in the
next version to allow it to be preprocessed out.
    A minor memory leak fixed in the error handling code. Also exclusion
handling code made more robust to blank or corrupt files being passed to it.

Admin:
    Tested quite thoroughly on a 32-bit Lazarus build.
    Stil allows configurations to be updated via UpgFetch.
    Test harnesses work.
    This version fairly robust and is nearly feature-complete.

Version 0.09. Tagged as 'ConfigUpd-0_09'