Summary:   Fixed bug in memory allocation check code.
Detail:
  Code has been added to fileswitch to detect whether there is sufficient
memory for a program to execute, and report a sensible error rather than
aborting in it's initialistion code.  This implicitly relied on R12 in
the error routines, which was being corrupted.  R12 register usage has now
replaced by other safe registers.

  This was exacerbated by incorrectly formed AIF headers, such as those
generated by at least some versions of the Easy C Linker, which don't have
correct entries for the size of the data areas of the file.

Admin:
  Tested on a RPC.  The system appears to work correctly, and malformed AIF
headers cause the error to be raised without other harm.


Version 2.64. Tagged as 'FileSwitch-2_64'