Summary:   Added new IOCTL reason codes and special field flags for sleeping devicefs.
Detail:
  DeviceFS has been enhanced with the ability to cooperate with Taskwindows
so as to allow multitasking to continue.  Any devicefs stream, if opened
with the special field entry 'sleep' will issue OS_UpCall 6 when it would
normally block.  A flag 'timeout' has also been added which will cause the
stream to issue a timeout error if it has slept for too long.

  These actions can be set using the generic IOCTLs 4 and 5, the former just
being a logical switch, the latter being the value in centiseconds for which
to sleep.

  There is also now a generic special field to specify non-blocking which may
be used as an alternative to the IOCTL.

Admin:
  Tested on a RiscPC.


Version 1.29. Tagged as 'HdrSrc-1_29'