Removed escape acknowledgement again.
This was a bad solution to the problem - acknowledging the escape says that
the module has dealt with it and the calling environment shouldn't take
escape actions. The C library is broken in that when the escape condition
is cleared/acknowledged, it still proceeds to call its SIGINT handler.
Compare this to the correct behaviour of the BASIC run-time; the problem
was originally noticed when calling *Debug, which you have to press Escape
to exit, and it acknowledges that escape, but C still throws SIGINT. When
the C library is fixed (as it shortly will be), if the Internet module
acknowledges escapes, socket calls will return EINTR but not actually
generate SIGINT.

The change was originally made for NFS, whose central code (in common with
most Unix code) retried calls returning EINTR. The problem was that the
escape condition remained unacknowledged so it got stuck in an infinite
loop. The correct fix will be to make sure that any supervisor mode code
such as NFS making blocking socket calls takes appropriate action when EINTR
is returned, or ensures that escape is disabled. Until it is fixed, NFS will
show lock-up symptoms again with this module. Applications will function as
expected, as the C run-time will call the SIGINT handler on exit from the
socket SWI (and acknowledge the escape) - if the signal handler returns the
foreground code can retry the call.

Version 5.45. Tagged as 'Internet-5_45'