Summary:   - Now uses no wimpslot so can just be run   - Now supports multiple telnet sessions
Detail:
  Previously the module had to be started with *desktop so that it was
started with the service call mechanism in order that the wimpslot could be
explicitly set.  Since the only memory required for C is for the Stack,
this is allocated in RMA.  Internal memory allocation is done from the RMA
explicitly.

  This enables the module to be run simply by double clicking on it rather
than having to start it's Task part separately, a technique used by ShareFS,
and maybe of use in other components.

  There is now support for multiple telnet sessions.  Previously only one was
supported at a time.  If debugit output is switched on, it will be displayed
in the last session opened.

Admin:
  Tested on a RiscPC.




Version 0.04. Tagged as 'TServer-0_04'