Summary:   - Fix to control transfers   - Clarification of root hub interrupts   - Error returns from USB Requests
Detail:
  Control transfers weren't properly counting the data portion across
multiple frames.  This is now properly counted.

  Root hub interrupts were only being turned on during reads.  They are now
on for the duration of the stream being open.

  If a control transfer fails during device call &80000000 (USB Request) then
the completion code is returned with an error message.

Admin:
  Tested on a DSL4000 and RiscPC.  It has been impossible to find a setting
of FSLargestDataPacket which is mutually compatible for full speed and low
speed devices.  Most low speed devices work with the setting &800 but some
mice appear not to, and only start working if the setting is changed to
&2000, however then printing stops working.

Version 0.05. Tagged as 'PHCIDriver-0_05'