Summary:   - Added command to add a remote subnet.
Detail:
  Freeway will interrogate an AUN server to discover subnets on which it
should broadcast.  However there was no other way to explicitly add another
subnet on which freeway should broadcast.

  The new *command FWAddnet will do this.  This allows ShareFS to talk to a
RISC OS machine on a different subet, e.g. between an STB and a RiscPC on the
engineering and house networks.  If the intervening router doesn't allow
directed broadcasts (as is quite often the case), you can specify a specific
host on the other network.

Admin:
  Tested on a DSL4000 and RiscPC on the house and engineering networks.


Version 0.31. Tagged as 'Freeway-0_31'