Summary:   - task closedown cleaned up   - media not present upcall trapped   - VDU processing much enhanced   - task quit fixed
Detail:
  The stack claimed from RMA is now released on Service_WimpCloseDown rather
than on exit.  This means it also gets released if the program exits
abnormally.  Service_WimpCloseDown also resets the task handle, so that if
the Wimp exits and restarts, TServer will restart as well, previously it was
failing to do this because it hadn't reset its task handle.

  There was a problem with media not present error boxes popping up, for
example when an attempt was made to access an unavailable share. This has
been fixed by sitting on UpCallV and cancelling it if the current task is one
of ours (determined by calling Wimp_ReadSysInfo).  This results in an
immediate response of 'shared disc not available' which is quite
satisfactory.

  Additional VDU Sequences now understood:

  6 and 21 turn on and off VDU output (control whether anything gets written
to the socket)

  17 sets foreground and background text colour

  30 homes the cursor

  31 tabs the cursor to position x, y.

  Multibyte VDU sequences other than these aren't eaten at the
moment.

  Task quit wan't being listened to.  Typing ShellCLI_TaskQuit in a session
will now quit it cleanly.  (you may like to alias this to something a little
easier to type!).

Admin:
  Tested extensively on RiscPC and STB3.



Version 0.10. Tagged as 'TServer-0_10'