Summary:   - now works in a taskwindow   - added toggle mode to repeatedly switch between channels
Detail:
  3 things stopped smallplay being useful in a taskwindow (e.g. telnet
session).  The screen wasn't blanked and it asked Videocontrol to change mode
and do a redraw.  The latter are now conditional upon the taskwindow status,
and if in a taskwindow, the first thing small play does is to generate a
transparent webpage and load it.

  The new toggle mode allows automatic switching between channels for te
purposes of testing.

Admin:
  Tested extensively on STB3.




Version 0.08. Tagged as 'SmallPlay-0_08'